ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
scikit-learn
python-docx
PyPDF2
python-multipart